+++ internal/clusterfeature/features/dns/manager.go
@@ -20,8 +20,8 @@ import (
 	"emperror.dev/errors"
 
 	"github.com/banzaicloud/pipeline/internal/clusterfeature"
+	"github.com/banzaicloud/pipeline/pkg/any"
 	"github.com/banzaicloud/pipeline/pkg/brn"
-	"github.com/banzaicloud/pipeline/pkg/opaque"
 )
 
 // FeatureManager implements the DNS feature manager
@@ -86,16 +86,16 @@ func (FeatureManager) ValidateSpec(ctx context.Context, spec clusterfeature.Feat
 
 // PrepareSpec makes certain preparations to the spec before it's sent to be applied
 func (m FeatureManager) PrepareSpec(ctx context.Context, clusterID uint, spec clusterfeature.FeatureSpec) (clusterfeature.FeatureSpec, error) {
-	defaulters := mapStringXform(map[string]opaque.Transformation{
-		"externalDns": mapStringDefaulter(map[string]opaque.Transformation{
+	defaulters := mapStringXform(map[string]any.Transformation{
+		"externalDns": mapStringDefaulter(map[string]any.Transformation{
 			"txtOwnerId": txtOwnerIDDefaulterXform(func() (string, error) {
 				return m.clusterUIDGetter.GetClusterUID(ctx, clusterID)
 			}),
 		}),
 	})
-	xform := mapStringXform(map[string]opaque.Transformation{
-		"externalDns": mapStringXform(map[string]opaque.Transformation{
-			"provider": mapStringXform(map[string]opaque.Transformation{
+	xform := mapStringXform(map[string]any.Transformation{
+		"externalDns": mapStringXform(map[string]any.Transformation{
+			"provider": mapStringXform(map[string]any.Transformation{
 				"secretId": secretBRNXform(func() (uint, error) {
 					return m.clusterOrgIDGetter.GetClusterOrgID(ctx, clusterID)
 				}),
@@ -103,7 +103,7 @@ func (m FeatureManager) PrepareSpec(ctx context.Context, clusterID uint, spec cl
 		}),
 	})
 
-	res, err := opaque.Compose(defaulters, xform).Transform(spec)
+	res, err := any.Compose(defaulters, xform).Transform(spec)
 	if err != nil {
 		return nil, errors.WrapIf(err, "failed to transform spec")
 	}
@@ -113,8 +113,8 @@ func (m FeatureManager) PrepareSpec(ctx context.Context, clusterID uint, spec cl
 	return nil, errors.Errorf("cannot cast type %T as type %T", res, spec)
 }
 
-func mapStringXform(transformations map[string]opaque.Transformation) opaque.Transformation {
-	return opaque.TransformationFunc(func(o interface{}) (interface{}, error) {
+func mapStringXform(transformations map[string]any.Transformation) any.Transformation {
+	return any.TransformationFunc(func(o interface{}) (interface{}, error) {
 		if m, ok := o.(map[string]interface{}); ok {
 			n := make(map[string]interface{}, len(m))
 			var errs error
@@ -133,8 +133,8 @@ func mapStringXform(transformations map[string]opaque.Transformation) opaque.Tra
 	})
 }
 
-func secretBRNXform(getOrgID func() (uint, error)) opaque.Transformation {
-	return opaque.TransformationFunc(func(secretObj interface{}) (interface{}, error) {
+func secretBRNXform(getOrgID func() (uint, error)) any.Transformation {
+	return any.TransformationFunc(func(secretObj interface{}) (interface{}, error) {
 		if secretStr, ok := secretObj.(string); ok {
 			orgID, err := getOrgID()
 			if err != nil {
@@ -153,8 +153,8 @@ func secretBRNXform(getOrgID func() (uint, error)) opaque.Transformation {
 	})
 }
 
-func mapStringDefaulter(trasformations map[string]opaque.Transformation) opaque.Transformation {
-	return opaque.TransformationFunc(func(o interface{}) (interface{}, error) {
+func mapStringDefaulter(trasformations map[string]any.Transformation) any.Transformation {
+	return any.TransformationFunc(func(o interface{}) (interface{}, error) {
 		if m, ok := o.(map[string]interface{}); ok {
 			n := make(map[string]interface{}, len(m))
 			for k, v := range m {
@@ -174,8 +174,8 @@ func mapStringDefaulter(trasformations map[string]opaque.Transformation) opaque.
 	})
 }
 
-func txtOwnerIDDefaulterXform(getClusterUID func() (string, error)) opaque.Transformation {
-	return opaque.TransformationFunc(func(txtOwnerIDObj interface{}) (interface{}, error) {
+func txtOwnerIDDefaulterXform(getClusterUID func() (string, error)) any.Transformation {
+	return any.TransformationFunc(func(txtOwnerIDObj interface{}) (interface{}, error) {
 		if txtOwnerIDStr, ok := txtOwnerIDObj.(string); ok && txtOwnerIDStr != "" {
 			return txtOwnerIDStr, nil
 		}
